+++ ignite/templates/field/datatype/string.go
@@ -12,6 +12,7 @@ import (
 var (
 	// DataString is a string data type definition.
 	DataString = DataType{
+		Name:                    String,
 		DataType:                func(string) string { return "string" },
 		CollectionsKeyValueName: func(string) string { return "collections.StringKey" },
 		DefaultTestValue:        "xyz",
@@ -40,6 +41,7 @@ var (
 
 	// DataStringSlice is a string array data type definition.
 	DataStringSlice = DataType{
+		Name:                    StringSlice,
 		DataType:                func(string) string { return "[]string" },
 		CollectionsKeyValueName: func(string) string { return collectionValueComment },
 		DefaultTestValue:        "abc,xyz",